+++ src/routes/changelog/articles/hp-rework.md
@@ -1,5 +1,5 @@
 ---
-date: 2022-01-23
+date: 2023-01-23
 title: HP Rework, videos, more!
 ---
