+++ src/routes/changelog/articles/report.md
@@ -1,0 +1,19 @@
+---
+date: 2023-01-31
+title: Report Button
+---
+
+**The Long requested report button has been added!**
+
+You can now report a video after you take a guess at it.
+
+I'm made aware of every report that comes in, and if a video gets a lot of reports, I will remove it from the rotation.
+
+## Some other changes I made
+
+- Form auto-fill-in and displayed rank now always use osu!standard rank
+- More rank exposure issues fixed
+
+Development on this is gonna slow down as I start getting into classes, but I still have a whole lot of ideas for where to go next, so stay tuned.
+
+As always, thanks so much for the support, it means more than you could ever imagine.
